pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-client"
version = "0.2.0"
authors = [
  { name="MCP Client Team", email="mcpclient@example.com" }
]
description = "An enhanced MCP client using Google Gemini for orchestration."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Utilities",
]
dependencies = [
    "google-generativeai>=0.5.0",
    "python-dotenv>=1.0.0",
    "mcp>=0.1.0",
    "click>=8.0",
    "rich>=13.0",
    "PyYAML>=6.0",
    "questionary>=2.0.0",
    "tiktoken>=0.6.0",
]

[project.urls]
"Homepage" = "https://github.com/mcp-client/mcp-client"
"Bug Tracker" = "https://github.com/mcp-client/mcp-client/issues"

# This makes the `mcp-client` command run the `cli` function in `main.py`
[project.scripts]
mcp-client = "mcp_client.main:cli"

[tool.setuptools.packages.find]
where = ["src"]
